Simple setup for generating emails
Relies on Handlebars, SCSSC, Emogrifier. Emails are produced by generating
CSS from the SCSS using SCSSC. The email html is compiled from a handlebars
template using Xamin Handlebars. The compiled html and css are sent to
Pelago Emogrifier, inlining the css into the html.

The final result is an email that is ready for distribution, which
is echo'd out for now.