Tracking the vendor folder
This is a controversial decision. On the one hand, the vendor folder
is a large set of dependencies that bloats the repo. On the other
hand, it is necessary for it to exist when deploying the plugin.

We are erring and the side of caution here and will track the vendor
folder for plugin portability.